+++ tests/tasks/check-coupled-files-spec.js
@@ -50,11 +50,11 @@ describe('task check-coupled-files', () => {
                 repo: 'bar',
                 number: 123,
                 body: 'Usually these filesets are changed together, but I detected some missing changes:\n\n' +
-                      '* in `[deps.json, deps.lock]` set there no change in these files: `[deps.lock]`\n' +
-                      '* in `[README.md, docs/README.html, docs/README.pdf]` set there no change ' +
+                      '* in `[deps.json, deps.lock]` set there\'s no change in these files: `[deps.lock]`\n' +
+                      '* in `[README.md, docs/README.html, docs/README.pdf]` set there\'s no change ' +
                       'in these files: `[README.md, docs/README.pdf]`\n\n' +
-                      'Please make sure that you didn\'t forget about something. ' +
-                       'If everything is all right, then sorry, my bad!'
+                      'Please make sure that you didn\'t forget anything. ' +
+                       'If everything is all right, then sorry - my bad!'
             });
 
             expect(logger.log).to.have.been.calledWithExactly('Posted info under pull request foo/bar#123');