+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Fix wrong `check-coupled-files` behaviour when none of files from fileSets were changed
+- Fix typos in `check-coupled-files` task wording
 
 ## [1.1.0] - 2017-12-04
 ### Added